pyproject.toml
[project]
name = "customgui"
version = "0.1.0"
description = "CustomUI 是一个开源的UI库，用于快速开发UI界面。"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easydict>=1.13",
    "pyside6>=6.10.1",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
windows-hardware = [
    "pyserial>=3.5; platform_system == 'Windows'",
    "hidapi>=0.14.0; platform_system == 'Windows'",
    "wmi>=1.5.1; platform_system == 'Windows'",
]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
